Refactor quantity fields to use integer type for consistency across models and migrations

+++ app/Filament/Resources/InventoryMovementResource.php
@@ -41,7 +41,7 @@ class InventoryMovementResource extends Resource
                     ->required(),
                 TextInput::make('qty')
                     ->label('Qty')
-                    ->numeric()
+                    ->integer()
                     ->required(),
                 TextInput::make('unit_cost')
                     ->label('Harga Modal')
